requirements.txt
fastapi==0.115.14
hub_sdk==0.0.20
Pillow==11.3.0
python-dotenv==1.1.1
ultralytics==8.3.158
uvicorn==0.35.0
